pyproject.toml
[project]
name = "learning-machine"
version = "0.1.0"
description = ""
authors = [
    {name = "Ekaterina Chernyshova",email = "newfatto@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "django (>=4.2,<5.0)",
    "djangorestframework (>=3.14,<4.0)",
    "psycopg2-binary (>=2.9.9,<3.0)",
    "pillow (>=10.0,<13.0)",
    "python-dotenv (>=1.0,<2.0)",
    "isort (>=7.0.0,<8.0.0)",
    "black (>=26.1.0,<27.0.0)",
    "django-filter (==23.5)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "coverage (>=7.13.4,<8.0.0)",
    "drf-yasg (>=1.21.15,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
